move TrainingFailedError out of _internal
