+++ python/ray/train/v2/_internal/exceptions.py
@@ -1,5 +1,5 @@
 import os
-from typing import Dict, List, Optional
+from typing import List, Optional
 
 from ray.train.v2._internal.constants import (
     DEFAULT_WORKER_GROUP_START_TIMEOUT_S,
@@ -40,18 +40,6 @@ class WorkerHealthCheckFailedError(RayTrainError):
 
     def __reduce__(self):
         return (self.__class__, (self._message, self.health_check_failure))
-
-
-class TrainingFailedError(RayTrainError):
-    """Exception raised by `<Framework>Trainer.fit()` when training fails."""
-
-    def __init__(self, error_message: str, worker_failures: Dict[int, Exception]):
-        super().__init__("Training failed due to worker errors:\n" + error_message)
-        self._error_message = error_message
-        self.worker_failures = worker_failures
-
-    def __reduce__(self):
-        return (self.__class__, (self._error_message, self.worker_failures))
 
 
 class WorkerGroupStartupTimeoutError(RayTrainError):
