+++ python/ray/train/v2/api/exceptions.py
@@ -1,0 +1,17 @@
+from typing import Dict
+
+from ray.train.v2._internal.exceptions import RayTrainError
+from ray.util.annotations import DeveloperAPI
+
+
+@DeveloperAPI
+class TrainingFailedError(RayTrainError):
+    """Exception raised by `<Framework>Trainer.fit()` when training fails."""
+
+    def __init__(self, error_message: str, worker_failures: Dict[int, Exception]):
+        super().__init__("Training failed due to worker errors:\n" + error_message)
+        self._error_message = error_message
+        self.worker_failures = worker_failures
+
+    def __reduce__(self):
+        return (self.__class__, (self._error_message, self.worker_failures))
