+++ python/ray/train/v2/tests/test_data_parallel_trainer.py
@@ -11,8 +11,8 @@ from ray.train.backend import Backend
 from ray.train.constants import RAY_CHDIR_TO_TRIAL_DIR, _get_ray_train_session_dir
 from ray.train.tests.util import create_dict_checkpoint
 from ray.train.v2._internal.constants import is_v2_enabled
-from ray.train.v2._internal.exceptions import TrainingFailedError
 from ray.train.v2.api.data_parallel_trainer import DataParallelTrainer
+from ray.train.v2.api.exceptions import TrainingFailedError
 from ray.train.v2.api.result import Result
 
 assert is_v2_enabled()
@@ -149,13 +149,9 @@ def test_error(tmp_path):
         scaling_config=ScalingConfig(num_workers=2),
         run_config=RunConfig(name="test", storage_path=str(tmp_path)),
     )
-    # with pytest.raises(TrainingFailedError) as exc_info:
-    try:
+    with pytest.raises(TrainingFailedError) as exc_info:
         trainer.fit()
-    except TrainingFailedError as e:
-        print(e)
-
-    # assert isinstance(exc_info.value.worker_failures[0], ValueError)
+        assert isinstance(exc_info.value.worker_failures[0], ValueError)
 
 
 @pytest.mark.parametrize("env_disabled", [True, False])
